Rest_Api_React_Assignment/CarRentingDjangoApi/requirements.txt
asgiref==3.5.2
Django==3.2.9
django-cleanup==6.0.0
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
Pillow==9.1.1
PyJWT==2.4.0
pytz==2022.1
sqlparse==0.4.2
